+++ CHANGELOG.md
@@ -23,7 +23,8 @@
   - ^ Executors removed entirely, it's functions moved to `Kutana` class.
   - ^ Removed `load_value` method.
   - ^ Renamed manager's `type` to `_type`.
-  - ^ Renamed manager's `get_background_coroutines` to `startup`
+  - ^ Renamed manager's `get_background_coroutines` to `startup`.
+  - ^ Added `Plugin.on_message` decorator.
   - Startup now accepts application.
   - Now kutana objects are referred as `app`.
   - Renamed Kutana's "process_update" to "process".